+++ laktory/cli/app.py
@@ -1,10 +1,11 @@
 import typer
+from laktory._version import VERSION
 
 APP_NAME = "laktory-cli"
 
 app = typer.Typer(
     pretty_exceptions_show_locals=False,
-    help="Laktory CLI to preview and deploy resources",
+    help=f"Laktory {VERSION} CLI to preview and deploy resources",
 )  # prevent display secret data